addRecordPage deleted, HomePage edited with new code, recordModal created
