+++ client/src/pages/HomePage/index.tsx
@@ -1,8 +1,19 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import '../assets/styles/index.css';
+import RecordModal from '../../components/recordModal';
+// import '../assets/styles/index.css';
 
 const HomePage = () => {
+  const [showModal, setShowModal] = useState(false);
+
+  const handleAddRecordClick = () => {
+    setShowModal(true);
+  };
+
+  const closeModal = () => {
+    setShowModal(false);
+  };
+
   return (
     <div className='home-page'>
       <header>
@@ -13,12 +24,17 @@ const HomePage = () => {
       </header>
       <main>
         <section className='actions'>
-          <Link to ='/add-record' className='btn'>Add a New Record Here</Link>
+          <button onClick={handleAddRecordClick} className='btn'>
+            Add a New Record Here
+          </button>
           <Link to ='/current-library' className='btn'>View Current Library Here</Link>
           <Link to ='/mood-music' className='btn'>Click here for the Mood Board</Link>
           <Link to='/wishlist' className='btn'>Rockin' Records Wishlist</Link>
         </section>
       </main>
+
+      {/* show modal when the state is true */}
+      {showModal && <RecordModal closeModal={closeModal} />}
     </div>
   );
 };
